pyproject.toml
[project]
name = "ds-star"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "pandas>=2.3.3",
    "scipy>=1.16.3",
    "tabulate>=0.9.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ruff>=0.14.6",
]
